feat(billing): metering migrations + amount_cents i64 (WIP-W3)
- 0050: llm_usage_events.usage_kind (+ index) for chat vs embedding exit metering
- 0051: billing_orders.amount_cents INT → BIGINT (matches i64 in memory)
- Persist usage_kind on insert_llm_usage_event
- Drop orphan usage_observer_impl (not wired; incomplete vs empty UsageObserver)

No separate staging gate: migrations are additive / type-widen; verify with
local migrate up/down on deploy.

+++ avrag-rs/crates/billing/src/service.rs
@@ -216,7 +216,7 @@ impl BillingService {
                         user_id,
                         &out_trade_no,
                         requested_plan,
-                        amount_cents as i32,
+                        amount_cents,
                     )
                     .await
                 {
